# Remove precaching of Django static files.