@rtibbles 2021-09-22
No - but that is disallowed by Django, so we should instead raise an error if it doesn't.

> It must end in a slash if set to a non-empty value.

https://docs.djangoproject.com/en/1.11/ref/settings/#std:setting-STATIC_URL